+++ Internet Archive Player tvOS/Views/TVDetail.swift
@@ -287,105 +287,7 @@ struct TVDetail: View {
     }
 }
 
-// MARK: - Supporting Views
-struct FlowingCollectionsView: View {
-    let collections: [Archive]
-    
-    var body: some View {
-        FlowLayout(spacing: 8) {
-            ForEach(Array(collections.prefix(12).enumerated()), id: \.element.id) { index, archive in
-                if let metadata = archive.metadata {
-                    HStack(spacing: 8) {
-                        // Collection thumbnail
-                        AsyncImage(url: metadata.iconUrl) { image in
-                            image
-                                .resizable()
-                                .aspectRatio(contentMode: .fill)
-                                .frame(width: 50, height: 50)
-                                .clipShape(RoundedRectangle(cornerRadius: 8))
-                        } placeholder: {
-                            RoundedRectangle(cornerRadius: 8)
-                                .fill(Color.gray.opacity(0.3))
-                                .frame(width: 50, height: 50)
-                                .overlay(
-                                    Image(systemName: "folder")
-                                        .font(.caption)
-                                        .opacity(0.5)
-                                )
-                        }
-                        
-                        // Collection title
-                        Text(metadata.archiveTitle ?? "Collection")
-                            .font(.caption)
-                            .opacity(0.9)
-                            .lineLimit(1)
-                    }
-                    .padding(.horizontal, 12)
-                    .padding(.vertical, 8)
-                    .background(Color.gray.opacity(0.2))
-                    .cornerRadius(10)
-                }
-            }
-        }
-    }
-}
-
-// Custom flow layout that wraps horizontally
-struct FlowLayout: Layout {
-    var spacing: CGFloat = 8
-    
-    func sizeThatFits(proposal: ProposedViewSize, subviews: Subviews, cache: inout ()) -> CGSize {
-        let result = FlowResult(
-            in: proposal.replacingUnspecifiedDimensions().width,
-            subviews: subviews,
-            spacing: spacing
-        )
-        return result.size
-    }
-    
-    func placeSubviews(in bounds: CGRect, proposal: ProposedViewSize, subviews: Subviews, cache: inout ()) {
-        let result = FlowResult(
-            in: bounds.width,
-            subviews: subviews,
-            spacing: spacing
-        )
-        for (index, subview) in subviews.enumerated() {
-            let position = result.positions[index]
-            subview.place(
-                at: CGPoint(x: bounds.minX + position.x, y: bounds.minY + position.y),
-                proposal: .unspecified
-            )
-        }
-    }
-    
-    struct FlowResult {
-        var size: CGSize = .zero
-        var positions: [CGPoint] = []
-        
-        init(in maxWidth: CGFloat, subviews: Subviews, spacing: CGFloat) {
-            var currentX: CGFloat = 0
-            var currentY: CGFloat = 0
-            var lineHeight: CGFloat = 0
-            
-            for subview in subviews {
-                let size = subview.sizeThatFits(.unspecified)
-                
-                if currentX + size.width > maxWidth && currentX > 0 {
-                    // Move to next line
-                    currentX = 0
-                    currentY += lineHeight + spacing
-                    lineHeight = 0
-                }
-                
-                positions.append(CGPoint(x: currentX, y: currentY))
-                lineHeight = max(lineHeight, size.height)
-                currentX += size.width + spacing
-            }
-            
-            self.size = CGSize(width: maxWidth, height: currentY + lineHeight)
-        }
-    }
-}
+// MARK: - Supporting Views (Private to TVDetail)
 
 private struct InfoBadge: View {
     let icon: String
@@ -567,486 +469,5 @@ extension TVDetail {
     }
 
 }
-// MARK: - Video Player with Placeholder
-struct VideoPlayerWithPlaceholder: View {
-    let videoURL: URL
-    let placeholderImageURL: URL?
-    
-    @State private var player: AVPlayer?
-    @State private var isLoading = true
-    
-    var body: some View {
-        ZStack {
-            // Placeholder while loading
-            if isLoading {
-                ZStack {
-                    Color.black
-                    
-                    if let imageURL = placeholderImageURL {
-                        AsyncImage(url: imageURL) { image in
-                            image
-                                .resizable()
-                                .aspectRatio(contentMode: .fit)
-                                .opacity(0.5)
-                        } placeholder: {
-                            ProgressView()
-                                .tint(.white)
-                        }
-                    } else {
-                        ProgressView()
-                            .tint(.white)
-                    }
-                    
-                    VStack(spacing: 12) {
-                        Spacer()
-                        ProgressView()
-                            .tint(.white)
-                            .scaleEffect(1.5)
-                        Text("Loading video...")
-                            .font(.caption)
-                            .foregroundColor(.white.opacity(0.8))
-                        Spacer()
-                    }
-                }
-                .ignoresSafeArea()
-            }
-            
-            // Video player
-            if let player = player {
-                VideoPlayer(player: player)
-                    .opacity(isLoading ? 0 : 1)
-                    .ignoresSafeArea()
-            }
-        }
-        .onAppear {
-            setupPlayer()
-        }
-        .onDisappear {
-            player?.pause()
-            player = nil
-        }
-    }
-    
-    private func setupPlayer() {
-        let playerItem = AVPlayerItem(url: videoURL)
-        let newPlayer = AVPlayer(playerItem: playerItem)
-        
-        // Observe when the player is ready to play
-        NotificationCenter.default.addObserver(
-            forName: .AVPlayerItemNewAccessLogEntry,
-            object: playerItem,
-            queue: .main
-        ) { _ in
-            withAnimation {
-                isLoading = false
-            }
-        }
-        
-        // Also check player status
-        playerItem.publisher(for: \.status)
-            .sink { status in
-                if status == .readyToPlay {
-                    withAnimation {
-                        isLoading = false
-                    }
-                }
-            }
-            .store(in: &cancellables)
-        
-        self.player = newPlayer
-        newPlayer.play()
-    }
-    
-    @State private var cancellables = Set<AnyCancellable>()
-}
 
-// MARK: - Audio Player View
-struct AudioPlayerView: View {
-    let audioFile: ArchiveFile
-    let artworkURL: URL?
-    let playlist: [ArchiveFile]? // Optional playlist
-    
-    @State private var player: AVPlayer?
-    @State private var isPlaying = false
-    @State private var currentTime: TimeInterval = 0
-    @State private var duration: TimeInterval = 0
-    @State private var timeObserverToken: Any?
-    @State private var currentTrackIndex: Int = 0
-    @State private var playlistFiles: [ArchiveFile] = []
-    @State private var seekIndicator: String? = nil // For showing +15s/-15s feedback
-    
-    // Seek interval in seconds (configurable)
-    private let seekInterval: Double = 15
-    
-    // Convenience initializer for single file playback
-    init(audioFile: ArchiveFile, artworkURL: URL?) {
-        self.audioFile = audioFile
-        self.artworkURL = artworkURL
-        self.playlist = nil
-    }
-    
-    // Full initializer with playlist support
-    init(audioFile: ArchiveFile, artworkURL: URL?, playlist: [ArchiveFile]?) {
-        self.audioFile = audioFile
-        self.artworkURL = artworkURL
-        self.playlist = playlist
-    }
-    
-    var currentTrack: ArchiveFile {
-        playlistFiles.isEmpty ? audioFile : playlistFiles[currentTrackIndex]
-    }
-    
-    var canGoNext: Bool {
-        currentTrackIndex < playlistFiles.count - 1
-    }
-    
-    var canGoPrevious: Bool {
-        currentTrackIndex > 0
-    }
-    
-    var body: some View {
-        ZStack {
-            // Background gradient
-            LinearGradient(
-                colors: [Color.black, Color.black.opacity(0.8)],
-                startPoint: .top,
-                endPoint: .bottom
-            )
-            .ignoresSafeArea()
-            
-            VStack(spacing: 50) {
-                Spacer()
-                
-                // Album artwork with seek indicator overlay
-                ZStack {
-                    if let imageURL = artworkURL {
-                        AsyncImage(url: imageURL) { image in
-                            image
-                                .resizable()
-                                .aspectRatio(contentMode: .fit)
-                                .frame(width: 600, height: 600)
-                                .cornerRadius(20)
-                                .shadow(color: .black.opacity(0.5), radius: 20, x: 0, y: 10)
-                        } placeholder: {
-                            RoundedRectangle(cornerRadius: 20)
-                                .fill(Color.gray.opacity(0.3))
-                                .frame(width: 600, height: 600)
-                                .overlay(
-                                    ProgressView()
-                                        .tint(.white)
-                                )
-                        }
-                    } else {
-                        RoundedRectangle(cornerRadius: 20)
-                            .fill(Color.gray.opacity(0.3))
-                            .frame(width: 600, height: 600)
-                            .overlay(
-                                Image(systemName: "music.note")
-                                    .font(.system(size: 120))
-                                    .foregroundColor(.white.opacity(0.5))
-                            )
-                    }
-                    
-                    // Seek indicator overlay
-                    if let indicator = seekIndicator {
-                        ZStack {
-                            RoundedRectangle(cornerRadius: 16)
-                                .fill(Color.black.opacity(0.75))
-                                .frame(width: 200, height: 100)
-                            
-                            Text(indicator)
-                                .font(.system(size: 40, weight: .bold))
-                                .foregroundColor(.white)
-                        }
-                        .transition(.scale.combined(with: .opacity))
-                    }
-                }
-                .frame(width: 600, height: 600)
-                
-                // Track info
-                VStack(spacing: 12) {
-                    Text(currentTrack.displayTitle)
-                        .font(.title2)
-                        .fontWeight(.bold)
-                        .foregroundColor(.white)
-                        .multilineTextAlignment(.center)
-                        .lineLimit(2)
-                    
-                    if let artist = currentTrack.artist ?? currentTrack.creator?.joined(separator: ", ") {
-                        Text(artist)
-                            .font(.title3)
-                            .foregroundColor(.white.opacity(0.8))
-                            .multilineTextAlignment(.center)
-                    }
-                    
-                    // Show track position if in playlist mode
-                    if !playlistFiles.isEmpty {
-                        Text("Track \(currentTrackIndex + 1) of \(playlistFiles.count)")
-                            .font(.caption)
-                            .foregroundColor(.white.opacity(0.6))
-                    }
-                }
-                
-                // Time progress bar
-                VStack(spacing: 8) {
-                    ProgressView(value: currentTime, total: max(duration, 0.1))
-                        .progressViewStyle(.linear)
-                        .tint(.white)
-                        .frame(width: 800)
-                    
-                    HStack {
-                        Text(formatTime(currentTime))
-                            .font(.caption)
-                            .foregroundColor(.white.opacity(0.7))
-                        
-                        Spacer()
-                        
-                        Text(formatTime(duration))
-                            .font(.caption)
-                            .foregroundColor(.white.opacity(0.7))
-                    }
-                    .frame(width: 800)
-                }
-                
-                // Playback controls
-                HStack(spacing: 80) {
-                    // Skip to previous track
-                    Button(action: playPrevious) {
-                        Image(systemName: "backward.end.fill")
-                            .font(.system(size: 44))
-                    }
-                    .disabled(!canGoPrevious)
-                    
-                    // Seek backward 15 seconds
-                    Button(action: seekBackward) {
-                        Image(systemName: "gobackward.\(Int(seekInterval))")
-                            .font(.system(size: 44))
-                    }
-                    
-                    // Play/Pause button
-                    Button(action: togglePlayPause) {
-                        Image(systemName: isPlaying ? "pause.circle.fill" : "play.circle.fill")
-                            .font(.system(size: 44))
-                    }
-                    
-                    // Seek forward 15 seconds
-                    Button(action: seekForward) {
-                        Image(systemName: "goforward.\(Int(seekInterval))")
-                            .font(.system(size: 44))
-                    }
-                    
-                    // Skip to next track
-                    Button(action: playNext) {
-                        Image(systemName: "forward.end.fill")
-                            .font(.system(size: 44))
-                    }
-                    .disabled(!canGoNext)
-                }
-                .padding(.bottom, 40)
-                
-                Spacer()
-            }
-            .padding(.horizontal, 90)
-            .padding(.top, 60)
-        }
-        .onAppear {
-            // Initialize playlist
-            if let playlist = playlist {
-                playlistFiles = playlist
-                // Find the starting index
-                if let index = playlist.firstIndex(where: { $0.name == audioFile.name }) {
-                    currentTrackIndex = index
-                }
-            } else {
-                playlistFiles = [audioFile]
-                currentTrackIndex = 0
-            }
-            setupPlayer()
-        }
-        .onDisappear {
-            cleanupPlayer()
-        }
-    }
-    
-    private func playNext() {
-        guard canGoNext else { return }
-        currentTrackIndex += 1
-        loadAndPlayCurrentTrack()
-    }
-    
-    private func playPrevious() {
-        guard canGoPrevious else { return }
-        currentTrackIndex -= 1
-        loadAndPlayCurrentTrack()
-    }
-    
-    private func seekForward() {
-        guard let player = player else { return }
-        let currentTime = player.currentTime()
-        let seekTime = CMTimeAdd(currentTime, CMTime(seconds: seekInterval, preferredTimescale: 1))
-        
-        // Make sure we don't seek past the end
-        if let duration = player.currentItem?.duration, seekTime < duration {
-            player.seek(to: seekTime, toleranceBefore: .zero, toleranceAfter: .zero)
-        } else if let duration = player.currentItem?.duration {
-            // If we'd go past the end, just go to the end
-            player.seek(to: duration, toleranceBefore: .zero, toleranceAfter: .zero)
-        }
-        
-        // Show visual feedback
-        showSeekIndicator("+\(Int(seekInterval))s")
-    }
-    
-    private func seekBackward() {
-        guard let player = player else { return }
-        let currentTime = player.currentTime()
-        let seekTime = CMTimeSubtract(currentTime, CMTime(seconds: seekInterval, preferredTimescale: 1))
-        
-        // Make sure we don't seek before the beginning
-        if seekTime > .zero {
-            player.seek(to: seekTime, toleranceBefore: .zero, toleranceAfter: .zero)
-        } else {
-            // If we'd go before the start, just go to the beginning
-            player.seek(to: .zero, toleranceBefore: .zero, toleranceAfter: .zero)
-        }
-        
-        // Show visual feedback
-        showSeekIndicator("-\(Int(seekInterval))s")
-    }
-    
-    private func showSeekIndicator(_ text: String) {
-        withAnimation(.spring(response: 0.3, dampingFraction: 0.7)) {
-            seekIndicator = text
-        }
-        
-        // Hide after 1 second
-        DispatchQueue.main.asyncAfter(deadline: .now() + 1.0) {
-            withAnimation(.easeOut(duration: 0.2)) {
-                seekIndicator = nil
-            }
-        }
-    }
-    
-    private func loadAndPlayCurrentTrack() {
-        // Clean up current player
-        cleanupPlayer()
-        
-        // Reset state
-        currentTime = 0
-        duration = 0
-        
-        // Setup new player for current track
-        setupPlayer()
-    }
-    
-    private func setupPlayer() {
-        guard let url = currentTrack.url else { return }
-        
-        let playerItem = AVPlayerItem(url: url)
-        let newPlayer = AVPlayer(playerItem: playerItem)
-        
-        // Observe player status to get duration
-        playerItem.publisher(for: \.status)
-            .sink { status in
-                if status == .readyToPlay {
-                    let durationSeconds = playerItem.duration.seconds
-                    if !durationSeconds.isNaN && durationSeconds.isFinite {
-                        duration = durationSeconds
-                    }
-                }
-            }
-            .store(in: &cancellables)
-        
-        // Observe when track finishes to auto-advance
-        NotificationCenter.default.publisher(for: .AVPlayerItemDidPlayToEndTime, object: playerItem)
-            .sink { _ in
-                if canGoNext {
-                    playNext()
-                } else {
-                    isPlaying = false
-                }
-            }
-            .store(in: &cancellables)
-        
-        // Add periodic time observer
-        let interval = CMTime(seconds: 0.5, preferredTimescale: CMTimeScale(NSEC_PER_SEC))
-        timeObserverToken = newPlayer.addPeriodicTimeObserver(forInterval: interval, queue: .main) { time in
-            let timeSeconds = time.seconds
-            if !timeSeconds.isNaN && timeSeconds.isFinite {
-                currentTime = timeSeconds
-            }
-        }
-        
-        self.player = newPlayer
-        newPlayer.play()
-        isPlaying = true
-    }
-    
-    private func cleanupPlayer() {
-        if let token = timeObserverToken {
-            player?.removeTimeObserver(token)
-            timeObserverToken = nil
-        }
-        player?.pause()
-        player = nil
-        cancellables.removeAll()
-    }
-    
-    private func togglePlayPause() {
-        guard let player = player else { return }
-        
-        if isPlaying {
-            player.pause()
-        } else {
-            player.play()
-        }
-        isPlaying.toggle()
-    }
-    
-    private func formatTime(_ time: TimeInterval) -> String {
-        guard time.isFinite && !time.isNaN else { return "0:00" }
-        let minutes = Int(time) / 60
-        let seconds = Int(time) % 60
-        return String(format: "%d:%02d", minutes, seconds)
-    }
-    
-    @State private var cancellables = Set<AnyCancellable>()
-}
-
-// MARK: - Player Time Observer (No longer needed, can be removed)
-class PlayerTimeObserver: ObservableObject {
-    var player: AVPlayer?
-    var timeObserver: Any?
-    var onTimeUpdate: ((TimeInterval, TimeInterval) -> Void)?
-    
-    init() {
-        startObserving()
-    }
-    
-    func startObserving() {
-        let interval = CMTime(seconds: 0.5, preferredTimescale: CMTimeScale(NSEC_PER_SEC))
-        timeObserver = player?.addPeriodicTimeObserver(forInterval: interval, queue: .main) { [weak self] time in
-            guard let self = self,
-                  let player = self.player else { return }
-            
-            let currentTime = time.seconds
-            let duration = player.currentItem?.duration.seconds ?? 0
-            
-            if !currentTime.isNaN && !duration.isNaN {
-                self.onTimeUpdate?(currentTime, duration)
-            }
-        }
-    }
-    
-    func stopObserving() {
-        if let timeObserver = timeObserver {
-            player?.removeTimeObserver(timeObserver)
-            self.timeObserver = nil
-        }
-    }
-    
-    deinit {
-        stopObserving()
-    }
-}
